Added the Penguin Prefab
- altered the worldcanvas of animal pens
- included penguin into spawner list
- also modified restaurant money gen. values

+++ SimulationProject/Assets/Scripts/BuildingFamily/Restaurant.cs
@@ -7,13 +7,13 @@ public class Restaurant : Building, ISpawnableObject
     //restaurant generates a random amount of money between
     //x and y
     [SerializeField]
-    protected float min;
+    protected int min;
     [SerializeField]
-    protected float max;
+    protected int max;
 
     protected override void GenerateMoney()
     {
-        float rand = Random.Range(min, max);
+        int rand = Random.Range(min, max);
         MoneyManager.SetCurrentIncome(rand);
     }
